+++ src/components/Switch/index.ts
@@ -1,1 +1,3 @@
+// Switch from copy-ui @ 2025-09-29.
+
 export { Switch } from "./Switch";