+++ src/components/Switch/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-09-28: Initial version.
+- 2024-09-29: Added controlled component support with `value` and `onChange`
+  props. Add `size` prop with 'sm', 'md', 'lg' options.